Downloads/MetaBitDepositBot-main/MetaBitDepositBot-mainV1/MetaBitDepositBot-main/requirements.txt
python-binance==1.0.19
python-telegram-bot[callback-data,job-queue]==20.7
python-dotenv==1.0.0
backoff==2.2.1
urllib3==2.0.7
requests==2.31.0
asyncio==3.4.3
setuptools>=65.5.1
web3==6.11.3
eth-account==0.10.0
aiosqlite
aiomysql
asyncpg==0.29.0
datadog
autopep8
psutil
flask==3.0.0
flask-login==0.6.3
aiohttp==3.9.1
pandas==2.1.4
plotly==5.18.0
plotly-express==0.4.1
psycopg2-binary==2.9.9
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
gunicorn==21.2.0
pytz==2023.3
xlsxwriter==3.2.0
openpyxl==3.1.2
tenacity==8.2.3
async-timeout==4.0.3
